Better Grid Selection Algorithm
Switched to using a 1D Kernel Density Estimator function with a Gaussian
Kernel. The implementation comes mostly from scikit learn's
documentation. By taking the local maxima of the Probability
Distribution Function, the concentraions where grid lines are likey to
be can be found.

There are some (not many) tests to ensure there isn't a regression in
the grid detection algoritm.

Next up is to write the algorithm which takes the selected nodes and the
estimated grid, computing the 2D array version of the text in the image.