Scale Annotations Based on Parent Size
It is now responsive at runtime. The size of the Konva stage is adjusted
as the window resizes.

Decided not to use a ResizeObserver the Annotations component takes up
the width of the window in the one place it is. Also, ResizeObserver
doesn't have native support and the pollyfill's look expensive (runtime).