Fixed OnMouseLeave Woes
Forgot to program in the one directional data flow way and was trying to
do things imperatively.

Now, every time a node is moused over, we just reset the selected nodes.
There is no concept of mousing out because we only need to handle the
case of mousing out of the top parent container.

Closes #11