Fixed Scrolling
The parent of the grid was centering the child. This is fine except when
the child became too small, it was centered with the overflow being
trimmed.

I couldn't find a way to get the children to overflow the flexbox.
Margin 0 auto centering, centers the child container without crushing it
down to the width of the parent.

Closes #4