Added GCloud Function
It works. The latency is surprisingly low. I guess my past experience
with functions may not be representative.

Decided to use GCloud functions as connecting to the database was a bit
easier than if using firebase.

At this point, the cloud resources for this project are getting kinda
messy and hard to keep track of.

There is a firebase project with firestore and analytics. There's a
surge hosting static files. There's a modeanalytics thing for dashboards
. There's a Cloud SQL database for ETLing data for analysis and storing
event data and there's a Cloud Function to get event data into the
database.

The firebase stuff lives in its own world, the mode stuff isn't worth
serializing. Maybe use terraform to deploy Cloud Functions and configure
Cloud SQL database?

Terraform is kinda annoying to setup and keep updated. Google's other
IAC suggestions kinda suck (fucking chef).

Let's just leave it as is. If I come back and forget how this whole
thing works, it is too complicated and I'll figure out how to
simplify/document it then.