Setup CORS
This endpoint need to be accessible from the actual site. I considered a
 few alts:

1. Cloudflare Workers as a proxy. This will add a performance floor on
the free plan of ~10ms.  Maybe this is really not a huge deal, but lets
keep the critical path clean.

Failed alts:

1. Redirect with Cloudflare Page Rules. Like proxying but easier.
Doesn't work unfortunately, redirecting doesn't endorse an Origin for
CORS.

2. Even localhost enforces CORS.

Will revisit CloudFlare if it is needed for SSR.